requirements.txt
torch
faiss-cpu
numpy
pandas
python-dotenv
openai
sentence-transformers
transformers
langchain
langchain-community
git+https://github.com/docling-ai/docling.git
git+https://github.com/docling-ai/docling-core.git
duckdb
langgraph
langchain-core
langchain-openai
opik
typing-extensions

